[irep2] Hoist the base-layout oracle into a shared unit
base_displacement and the three helpers it picks between were file-local
statics in clang_c_adjust_expr.cpp. The IREP2 adjust pass needs the same
answer, and taking the offset from anywhere else -- clang's record layout
in particular -- is the mistake #3894 records, so there has to be exactly
one oracle rather than a second implementation.

Pure move: only base_displacement is exposed.
